Remove try catch statements
I believe error handling should be done outside of these logical functions.
The functions should be correctly implemented, as if given a string,
it should always work.
This means if we find some bug given the correct input, the method
should be improved, but if the method is called with the wrong input,
this error handling should happen outside the function.
Up for discussion.