Add useQuery to CreateProjectDialog
Also simplified the component by reducing communication with ProjectList.
This could perhaps been a seperate commit. However, this was developed
side by side and is a side effect of React query. Would be a hassle to
seperate these commits, be merciful with me.
Also our very first test broke. It is difficult to unittest this when
depending on a useQueryClient wrapper. Can be reimplemented later when
we want more end-to-end/userstory kind of tests.