Force missing input values to 0
Users expected missing input to be 0 in simulations, which
instead caused it to crash. Now if a input field is empty, it will
be evaluated as 0 and have 0 as placeholder.
At first i tried having value to be 0 if missing value, but it
was a bit annoying having to delete the 0 before you can change
the field to something else, therefore I went with placeholder 0
and no value actual value in the field if 0, while 0 is still
saved to the formconfig.