Remove unneccessary divs
For performance and code smell purposes, divs should not be
overused as they cause complexity for the DOM.
Also, generally, I believe components should describe it's
contents, while the styling may vary based on its use, which
should be decided one level up.
For example, project list needs a certain width to fit nicely
on the page. But this should be defined in the page where project
list is rendered, as this style could change for other pages
containing projectlist in the future.

Change my mind!