Validate in frontend using zod
This commit adds zod and replaces dto with zod-based schema. To make
this change easier, `apiRequest` now has a `responseModel` "kwarg" which
is a zod schema and is validated. Errors are logged. Additionally, it's
now not possible to directly obtain a JSON from `apiRequest` unless you
go via `responseReturn: true` which gives a raw `Response` instance.